Adds orc_utils rosjava project with decompiled source from previous orc jars
There was some inconsistency between the OrcStatus class in orc-0.0.jar and
the one in uORCInterface-0.0.jar. We wanted the OrcStatus included with
uORCInterface-0.0.jar, but we wanted the extra classes in orc-0.0.jar, so I
decompiled both jars, and stuck together the appropriate bits. I actually
ended up changing uORCInterface so that it could rely on orc-0.0.jar
untouched, but also had to change some source, since this was compiling
with java 6, instead of java 7, as the JARs were originally compiled with
(it appears).